*** handle configuration and history with C code ***
This is a major change that may lead to a complete C application.

Deleted files:
- prefs.cc
- prefs.h
- history.cc
- history.h

New files:
- config_prefs.c
- config_prefs.h
- history.c
- history.h

gmrun was mostly broken, I'm not sure when the app became so broken
but it's better now.

The code is more readable and more stable.

Various bugs were fixed, it would take many lines to explain the changes,
they were just so dramatic